# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/billing/plans": {
      "get": {
        "description": "Plans are rows, resolved server-side. The app renders what it is told — a price change or a new tier is an insert, not a release.",
        "operationId": "Billing_plans",
        "parameters": [
          {
            "name": "payer",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "family",
                "dispatchOrganization"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionPlanDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "The plan catalogue",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/account": {
      "get": {
        "operationId": "Billing_account",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "The caller's household subscription",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/subscribe": {
      "post": {
        "description": "Refused if one is already live — changing plan is `change-interval`, and anything else is two intentions sharing an endpoint.",
        "operationId": "Billing_subscribe",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opaque, client-generated, 8–200 characters. Repeating a request with the same key returns the first response rather than performing it again. Same key with a different body is refused. Honoured for 24 hours.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start a household subscription",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/change-interval": {
      "post": {
        "description": "The unused remainder of the current period is credited and a fresh period starts today. Annual → monthly leaves a credit carried against renewals, not a refund.",
        "operationId": "Billing_changeInterval",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeIntervalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Move between monthly and annual",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/invoices": {
      "get": {
        "description": "Newest first. Line items are read back from the invoice, not recomputed — a superseded plan must not reprint last March at this March’s prices.",
        "operationId": "Billing_invoices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "What this household has been billed",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/payment-method": {
      "post": {
        "description": "Takes a token the client obtained directly from the processor. No endpoint in this system accepts a card number — see ADR-0006.",
        "operationId": "Billing_attachPaymentMethod",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opaque, client-generated, 8–200 characters. Repeating a request with the same key returns the first response rather than performing it again. Same key with a different body is refused. Honoured for 24 hours.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachPaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethodDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Put a card on file",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/payment-method/{id}": {
      "delete": {
        "description": "The row is kept and marked detached, so a payment made months ago still names the card that made it. Removing the last card is allowed.",
        "operationId": "Billing_detachPaymentMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Take a card off file",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/invoices/{id}/pay": {
      "post": {
        "description": "For the moment after a declined card is replaced: waiting a day for the scheduled retry, while the screen still says the payment failed, reads as the update not having worked.",
        "operationId": "Billing_payInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Charge an open invoice now",
        "tags": [
          "billing"
        ]
      }
    },
    "/billing/cancel": {
      "post": {
        "description": "Not a refund and not an immediate switch-off: a family part-way through a booked month keeps live tracking for the rides they have already arranged.",
        "operationId": "Billing_cancel",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cancel at the end of the paid period",
        "tags": [
          "billing"
        ]
      }
    },
    "/organizations": {
      "get": {
        "operationId": "Organizations_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Organisations the caller belongs to",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing": {
      "get": {
        "operationId": "Organizations_organizationBilling",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "An operator's subscription",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/subscribe": {
      "post": {
        "description": "Priced at the drivers actually on the road — an operator cannot subscribe at five seats and run twenty.",
        "operationId": "Organizations_subscribeOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opaque, client-generated, 8–200 characters. Repeating a request with the same key returns the first response rather than performing it again. Same key with a different body is refused. Honoured for 24 hours.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start an operator subscription",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/change-interval": {
      "post": {
        "operationId": "Organizations_changeOrganizationInterval",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeIntervalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Move an operator between monthly and annual",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/cancel": {
      "post": {
        "operationId": "Organizations_cancelOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingAccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cancel an operator subscription at period end",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/invoices": {
      "get": {
        "operationId": "Organizations_organizationInvoices",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "What this operator has been billed",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/payment-method": {
      "post": {
        "operationId": "Organizations_attachOrganizationPaymentMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Opaque, client-generated, 8–200 characters. Repeating a request with the same key returns the first response rather than performing it again. Same key with a different body is refused. Honoured for 24 hours.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachPaymentMethodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentMethodDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Put a card on an operator's account",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/payment-method/{paymentMethodId}": {
      "delete": {
        "operationId": "Organizations_detachOrganizationPaymentMethod",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paymentMethodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Take a card off an operator's account",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/billing/invoices/{invoiceId}/pay": {
      "post": {
        "operationId": "Organizations_payOrganizationInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invoiceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Charge an operator's open invoice now",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/seats": {
      "get": {
        "description": "The audit trail an invoice line is answerable from: without it, \"why were we billed for eleven drivers\" can only be answered from a driver table that has since changed.",
        "operationId": "Organizations_seats",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationSeatsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Drivers, seats and the ledger behind them",
        "tags": [
          "organizations"
        ]
      }
    },
    "/organizations/{id}/seats/quote": {
      "get": {
        "operationId": "Organizations_quote",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "planCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "interval",
            "required": true,
            "in": "query",
            "schema": {
              "enum": [
                "monthly",
                "annual"
              ],
              "type": "string"
            }
          },
          {
            "name": "seats",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionQuoteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "What a given driver count would cost",
        "tags": [
          "organizations"
        ]
      }
    },
...
  },
  "info": {
    "title": "CareBridge API",
    "description": "Coordination for an older relative’s medical appointments and the transport to them.\n\nTwo conventions hold across every endpoint:\n\n- **Errors** use one envelope: `{ error: { code, message, correlationId, field? } }`.\n- **`404 not_found_or_forbidden` is deliberately ambiguous.** \"No such record\" and\n  \"not yours\" return an identical response, so the API cannot be used to probe for\n  the existence of a patient, a ride or an account.\n\nThis service stores no diagnoses, medications, clinical notes or dates of birth.\nSee `docs/FOUNDATION.md` §9.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API, direct"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local, through the nginx proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token from POST /auth/login. Carries the user id and nothing else — roles, organisations and patient grants are resolved server-side per request so revocation takes effect on the next call."
      }
    },
    "schemas": {
...
      "BillingInterval": {
        "type": "string",
        "enum": [
          "monthly",
          "annual"
        ]
      },
...
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "description": "Plan code, e.g. \"family-standard\" or \"dispatch-core\"."
          },
          "interval": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BillingInterval"
              }
            ]
          }
        },
        "required": [
          "planCode",
          "interval"
        ]
      },
      "ChangeIntervalDto": {
        "type": "object",
        "properties": {
          "interval": {
            "description": "The unused remainder of the current period is credited; a fresh period starts today. Annual → monthly produces a credit carried forward, not a refund.",
            "allOf": [
              {
                "$ref": "#/components/schemas/BillingInterval"
              }
            ]
          }
        },
        "required": [
          "interval"
        ]
      },
...
    }
  }
}